Check length of array before reducing
Sometimes, around line 1024, the array named v will be empty,
and trying to reduce an empty array will throw an error. This
is just a quick fix for just that. Here, we check that the
length of v is bigger than 0, and only reduce the array if it
has any elements.